+++ Source/Quarry/Quarry.csproj
@@ -52,6 +52,7 @@
     <Compile Include="Static\Extensions.cs" />
     <Compile Include="Orders\Designator_ReclaimSoil.cs" />
     <Compile Include="AI\JobDriver_MineQuarry.cs" />
+    <Compile Include="Static\HarmonyPatches.cs" />
     <Compile Include="Static\QuarryDefOf.cs" />
     <Compile Include="Buildings\Building_Quarry.cs" />
     <Compile Include="PlaceWorkers\PlaceWorker_Quarry.cs" />
